+++ tests/models/awards/CoreAward-test.js
@@ -6,6 +6,9 @@
 import CoreAward from 'models/v2/awards/CoreAward';
 
 const awardData = {
+    category: null,
+    startDate: '1989-01-02',
+    endDate: '1999-12-31',
     subawardTotal: '12004.75'
 };
 
@@ -13,6 +16,13 @@ const award = Object.create(CoreAward);
 award.populateCore(awardData);
 
 describe('Core Award getter functions', () => {
+    it('should use IDV when the category is null', () => {
+        expect(award.category).toEqual('idv');
+    });
+    it('should properly format the start and end dates', () => {
+        expect(award.startDate).toEqual('01/02/1989');
+        expect(award.endDate).toEqual('12/31/1999');
+    });
     it('should format the subaward total', () => {
         expect(award.subawardTotal).toEqual('$12,005');
     });
