+++ tests/models/awardsV2/CoreExecutiveDetails-test.js
@@ -3,18 +3,17 @@
  * Created by David Trinh 10/6/18
  */
 
-import CoreExecutiveDetails from 'models/v2/awards/CoreExecutiveDetails';
-import { mockLoan } from './mockAwardApi';
+import CoreExecutiveDetails from 'models/v2/awardsV2/CoreExecutiveDetails';
+import { mockContract } from './mockAwardApi';
 
 const executiveDetails = Object.create(CoreExecutiveDetails);
-executiveDetails.populateCore(mockLoan.executive_details);
+executiveDetails.populateCore(mockContract.executive_details);
 
 describe('CoreExecutiveDetails', () => {
     it('should parse executive compensation', () => {
         expect(executiveDetails.officers).toEqual({
             officer1: "John Doe - $12,132",
-            officer2: "Jake Doe - $0",
-            officer3: "--"
+            officer2: "Jake Doe - $0"
         });
     });
     it('should parse executive compensation even without data', () => {
